requirements.txt
torchdiffeq-0.2.3
torchsummary-1.5.1
pyyaml-6.0
einops-0.6.0